Move util.py to utils
