+++ python/ray/tune/utils/__init__.py
@@ -1,0 +1,9 @@
+from ray.tune.utils.util import (deep_update, flatten_dict, get_pinned_object,
+                                 merge_dicts, pin_in_object_store, UtilMonitor,
+                                 validate_save_restore, warn_if_slow)
+
+__all__ = [
+    "deep_update", "flatten_dict", "get_pinned_object", "merge_dicts",
+    "pin_in_object_store", "UtilMonitor", "validate_save_restore",
+    "warn_if_slow"
+]
